requirements.txt
alembic==1.12.1
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
flake8==6.1.0
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==2.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==1.24.3
Werkzeug==2.3.7
